+++ tests/adapters/test_base.py
@@ -28,20 +28,27 @@ def test_adapter_get_columns():
 
 def test_adapter_get_data():
     adapter = FakeAdapter()
-    data = adapter.get_data({})
+
+    data = adapter.get_data({}, [])
     assert list(data) == [
         {"rowid": 0, "name": "Alice", "age": 20, "pets": 0},
         {"rowid": 1, "name": "Bob", "age": 23, "pets": 3},
     ]
 
-    data = adapter.get_data({"name": Equal("Alice")})
+    data = adapter.get_data({"name": Equal("Alice")}, [])
     assert list(data) == [
         {"rowid": 0, "name": "Alice", "age": 20, "pets": 0},
     ]
 
-    data = adapter.get_data({"age": Range(20, None, False, False)})
+    data = adapter.get_data({"age": Range(20, None, False, False)}, [])
     assert list(data) == [
         {"rowid": 1, "name": "Bob", "age": 23, "pets": 3},
+    ]
+
+    data = adapter.get_data({}, [("age", Order.DESCENDING)])
+    assert list(data) == [
+        {"rowid": 1, "name": "Bob", "age": 23, "pets": 3},
+        {"rowid": 0, "name": "Alice", "age": 20, "pets": 0},
     ]
 
 
@@ -49,14 +56,14 @@ def test_adapter_manipulate_rows():
     adapter = FakeAdapter()
 
     adapter.insert_row({"rowid": None, "name": "Charlie", "age": 6, "pets": 1})
-    data = adapter.get_data({})
+    data = adapter.get_data({}, [])
     assert list(data) == [
         {"rowid": 0, "name": "Alice", "age": 20, "pets": 0},
         {"rowid": 1, "name": "Bob", "age": 23, "pets": 3},
         {"rowid": 2, "name": "Charlie", "age": 6, "pets": 1},
     ]
     adapter.insert_row({"rowid": 4, "name": "Dani", "age": 40, "pets": 2})
-    data = adapter.get_data({})
+    data = adapter.get_data({}, [])
     assert list(data) == [
         {"rowid": 0, "name": "Alice", "age": 20, "pets": 0},
         {"rowid": 1, "name": "Bob", "age": 23, "pets": 3},
@@ -65,7 +72,7 @@ def test_adapter_manipulate_rows():
     ]
 
     adapter.delete_row(0)
-    data = adapter.get_data({})
+    data = adapter.get_data({}, [])
     assert list(data) == [
         {"rowid": 1, "name": "Bob", "age": 23, "pets": 3},
         {"rowid": 2, "name": "Charlie", "age": 6, "pets": 1},
@@ -73,7 +80,7 @@ def test_adapter_manipulate_rows():
     ]
 
     adapter.update_row(1, {"rowid": 1, "name": "Bob", "age": 24, "pets": 4})
-    data = adapter.get_data({})
+    data = adapter.get_data({}, [])
     assert list(data) == [
         {"rowid": 2, "name": "Charlie", "age": 6, "pets": 1},
         {"rowid": 4, "name": "Dani", "age": 40, "pets": 2},